# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST de gestion de bibliothèque universitaire.",
        "title": "Bibliothèque API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/auteurs": {
            "get": {
                "description": "Retourne la liste des auteurs disponibles.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ouvrages"
                ],
                "summary": "Lister les auteurs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Auteur"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/emplacements": {
            "get": {
                "description": "Retourne la liste des emplacements disponibles.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ouvrages"
                ],
                "summary": "Lister les emplacements",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repositories.EmplacementResume"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/livres": {
            "post": {
                "description": "Cree un nouveau livre avec auteur existant ou nouvel auteur.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ouvrages"
                ],
                "summary": "Creer un livre",
                "parameters": [
                    {
                        "description": "Donnees du livre",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreerLivreRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/revues": {
            "post": {
                "description": "Cree une nouvelle revue.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ouvrages"
                ],
                "summary": "Creer une revue",
                "parameters": [
                    {
                        "description": "Donnees de la revue",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreerRevueRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/utilisateurs/{id}/caution": {
            "get": {
                "description": "Retourne le solde de caution disponible et la caution totale.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Consulter la caution d'un utilisateur",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID utilisateur",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repositories.CautionInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Met a jour la caution totale d'un utilisateur et recalcule son solde.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Mettre a jour la caution totale",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID utilisateur",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nouvelle caution",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateCautionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repositories.CautionInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controllers.CreerLivreRequest": {
            "type": "object",
            "properties": {
                "auteur_id": {
                    "type": "integer"
                },
                "auteur_nom": {
                    "type": "string"
                },
                "auteur_prenom": {
                    "type": "string"
                },
                "caution": {
                    "type": "number"
                },
                "emplacement_id": {
                    "type": "integer"
                },
                "isbn": {
                    "type": "string"
                },
                "titre": {
                    "type": "string"
                }
            }
        },
        "controllers.CreerRevueRequest": {
            "type": "object",
            "properties": {
                "caution": {
                    "type": "number"
                },
                "date_parution": {
                    "type": "string"
                },
                "emplacement_id": {
                    "type": "integer"
                },
                "numero": {
                    "type": "integer"
                },
                "titre": {
                    "type": "string"
                }
            }
        },
...
        "controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "erreur": {
                    "type": "string"
                }
            }
        },
...
        "controllers.IDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.UpdateCautionRequest": {
            "type": "object",
            "properties": {
                "caution_totale": {
                    "type": "number"
                }
            }
        },
        "models.Auteur": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nom": {
                    "type": "string"
                },
                "prenom": {
                    "type": "string"
                }
            }
        },
...
        "repositories.CautionInfo": {
            "type": "object",
            "properties": {
                "caution_totale": {
                    "type": "number"
                },
                "solde_caution": {
                    "type": "number"
                }
            }
        },
        "repositories.EmplacementResume": {
            "type": "object",
            "properties": {
                "categorie_nom": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "niveau": {
                    "type": "integer"
                },
                "numero_etagere": {
                    "type": "integer"
                },
                "numero_travee": {
                    "type": "integer"
                }
            }
        },
...
    }
}